[Frontend] add iquest reasoning parser to support iquest chat template
